WIP: refactor to remove hyperopt sampler
Migrating sampler functionality to Executor
